UnindexedFileIterator: mtime is quint32 not quint64

+++ src/file/unindexedfileiterator.cpp
@@ -85,7 +85,7 @@ bool UnIndexedFileIterator::shouldIndex(const QString& filePath, const QString& 
         return true;
     }
 
-    quint64 mTime = m_transaction->documentMTime(fileId);
+    quint32 mTime = m_transaction->documentMTime(fileId);
 
     // A folders mtime is updated when a new file is added / removed / renamed
     // we don't really need to reindex a folder when that happens
@@ -97,6 +97,5 @@ bool UnIndexedFileIterator::shouldIndex(const QString& filePath, const QString& 
     if (mTime != fileInfo.lastModified().toTime_t()) {
         return true;
     }
-
     return false;
 }
